+++ src/lib/typed-storage-key.ts
@@ -7,7 +7,7 @@ export class TypedStorageKey<T> {
 
     public get key(): string { return this._key; }
     public get type(): { new(): T } { return this._t; }
-    public get typeName(): string { return typeof this.type === "string" ? "String" : this.type.name; }
+    public get typeName(): string { return this.type.name; }
 
     public toString(): string {
         return this.key;